+++ src/Common/Class/GameScene/PhysicEngine/Force/set.cpp
@@ -10,12 +10,12 @@ void    CS_Force::setAccelerationY(float aYsource)
     aY = Tools->transformHeight(aYsource);
 }
 
-void    CS_Force::setAccelerationXPixel(int aXsource)
+void    CS_Force::setAccelerationXPixel(float aXsource)
 {
     aX = aXsource;
 }
 
-void    CS_Force::setAccelerationYPixel(int aYsource)
+void    CS_Force::setAccelerationYPixel(float aYsource)
 {
     aY = aYsource;
 }
