+++ src/Common/Class/GameScene/PhysicEngine/Force/use.cpp
@@ -1,6 +1,6 @@
 # include "gameScene.h"
 
-void    CS_Force::usePhysic(int& vXDest, int& vYDest, float deltaT)
+void    CS_Force::usePhysic(float& vXDest, float& vYDest, int deltaT)
 {
     vXDest += aX * deltaT;
     vYDest += aY * deltaT;
