+++ src/Common/Class/GameScene/PhysicEngine/PersonalPhysic/update.cpp
@@ -1,11 +1,11 @@
 # include "gameScene.h"
 
-void    CS_PersonalPhysic::updateSpeed(float deltaT)
+void    CS_PersonalPhysic::updateSpeed(int deltaT)
 {
     speed.updateSpeed(WorldPhysics->QueryGravity(), deltaT);
 }
 
-void    CS_PersonalPhysic::updatePosition(float deltaT)
+void    CS_PersonalPhysic::updatePosition(int deltaT)
 {
     position.updatePosition(speed, deltaT);
 }
